requirements.txt
comtypes==1.1.7
numpy==1.18.5
pygame==2.0.3
pykinect2==0.1.0
Pillow==6.2.2